cumulative counts in HistVal, cap input for Val
Adding a CumCount field to HistVal does not require additional
computation for any of the methods provided today and has real
uses. Switching Percentile to be a method is not possible because
that would require freezing the total count, which is clumsy at
best.

Additionally, allow Val to take values higher than HighestTrackable
as arguments. Previously this would result in a panic, but
effectively capping it to be the HighestTrackable provides safe,
sensible behavior.